+++ io_context/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>io_context</name>
+  <name>boost_io_context</name>
   <version>1.2.0</version>
   <description>A library to write Synchronous and Asynchronous networking applications</description>
   <maintainer email="whitleysoftwareservices@gmail.com">Joshua Whitley</maintainer>
