+++ io_context/test/test_io_context.cpp
@@ -19,7 +19,7 @@
 #include <vector>
 #include <memory>
 
-#include "io_context/io_context.hpp"
+#include "boost_io_context/io_context.hpp"
 
 using drivers::common::IoContext;
