+++ serial_driver/include/boost_serial_driver/serial_driver.hpp
@@ -18,8 +18,8 @@
 #include <memory>
 #include <string>
 
-#include "io_context/io_context.hpp"
-#include "serial_driver/serial_port.hpp"
+#include "boost_io_context/io_context.hpp"
+#include "boost_serial_driver/serial_port.hpp"
 
 namespace drivers
 {
