+++ tcp_driver/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-    <name>tcp_driver</name>
+    <name>boost_tcp_driver</name>
     <version>0.0.1</version>
     <description>A library to write Synchronous and Asynchronous networking applications, ROS and ROS2 nodes
     </description>
