+++ udp_driver/include/boost_udp_driver/udp_receiver_node.hpp
@@ -12,20 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef UDP_DRIVER__UDP_SENDER_NODE_HPP_
-#define UDP_DRIVER__UDP_SENDER_NODE_HPP_
+#ifndef UDP_DRIVER__UDP_RECEIVER_NODE_HPP_
+#define UDP_DRIVER__UDP_RECEIVER_NODE_HPP_
 
-#include "udp_driver/udp_driver.hpp"
+#include "boost_udp_driver/udp_driver.hpp"
 
 #include <chrono>
 #include <memory>
 #include <string>
+#include <vector>
 
 #include <rclcpp/rclcpp.hpp>
 #include <rclcpp_lifecycle/lifecycle_node.hpp>
 #include <lifecycle_msgs/msg/state.hpp>
 
-#include "msg_converters/converters.hpp"
+#include "boost_msg_converters/converters.hpp"
 
 namespace lc = rclcpp_lifecycle;
 using LNI = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface;
@@ -35,24 +36,24 @@ namespace drivers
 namespace udp_driver
 {
 
-/// \brief UdpSenderNode class which can send UDP datagrams
-class UdpSenderNode final
+/// \brief UdpReceiverNode class which can receive UDP datagrams
+class UdpReceiverNode final
   : public lc::LifecycleNode
 {
 public:
   /// \brief Default constructor
   /// \param[in] options Options for the node
-  explicit UdpSenderNode(const rclcpp::NodeOptions & options);
+  explicit UdpReceiverNode(const rclcpp::NodeOptions & options);
 
   /// \brief Constructor which accepts IoContext
   /// \param[in] options Options for the node
   /// \param[in] ctx A shared IoContext
-  UdpSenderNode(
-    const rclcpp::NodeOptions & options,
+  UdpReceiverNode(
+    const rclcpp::NodeOptions &,
     const drivers::common::IoContext & ctx);
 
   /// \brief Destructor - required to manage owned IoContext
-  ~UdpSenderNode();
+  ~UdpReceiverNode();
 
   /// \brief Callback from transition to "configuring" state.
   /// \param[in] state The current state that the node is in.
@@ -74,8 +75,8 @@ public:
   /// \param[in] state The current state that the node is in.
   LNI::CallbackReturn on_shutdown(const lc::State & state) override;
 
-  /// \brief Callback for sending a UDP datagram
-  void subscriber_callback(udp_msgs::msg::UdpPacket::SharedPtr msg);
+  /// \breif Callback for receiving a UDP datagram
+  void receiver_callback(const std::vector<uint8_t> & buffer);
 
 private:
   void get_params();
@@ -84,10 +85,10 @@ private:
   std::string m_ip{};
   uint16_t m_port{};
   std::unique_ptr<UdpDriver> m_udp_driver;
-  rclcpp::Subscription<udp_msgs::msg::UdpPacket>::SharedPtr m_subscriber;
-};  // class UdpSenderNode
+  lc::LifecyclePublisher<udp_msgs::msg::UdpPacket>::SharedPtr m_publisher;
+};  // class UdpReceiverNode
 
 }  // namespace udp_driver
 }  // namespace drivers
 
-#endif  // UDP_DRIVER__UDP_SENDER_NODE_HPP_
+#endif  // UDP_DRIVER__UDP_RECEIVER_NODE_HPP_
