+++ udp_driver/include/boost_udp_driver/udp_sender_node.hpp
@@ -12,49 +12,47 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef SERIAL_DRIVER__SERIAL_BRIDGE_NODE_HPP_
-#define SERIAL_DRIVER__SERIAL_BRIDGE_NODE_HPP_
+#ifndef UDP_DRIVER__UDP_SENDER_NODE_HPP_
+#define UDP_DRIVER__UDP_SENDER_NODE_HPP_
 
-#include "serial_driver/serial_driver.hpp"
+#include "boost_udp_driver/udp_driver.hpp"
 
 #include <chrono>
 #include <memory>
 #include <string>
-#include <vector>
 
 #include <rclcpp/rclcpp.hpp>
 #include <rclcpp_lifecycle/lifecycle_node.hpp>
 #include <lifecycle_msgs/msg/state.hpp>
 
-#include "msg_converters/converters.hpp"
+#include "boost_msg_converters/converters.hpp"
 
 namespace lc = rclcpp_lifecycle;
 using LNI = rclcpp_lifecycle::node_interfaces::LifecycleNodeInterface;
-using std_msgs::msg::UInt8MultiArray;
 
 namespace drivers
 {
-namespace serial_driver
+namespace udp_driver
 {
 
-/// \brief SerialBridgeNode class which can send and receive serial data
-class SerialBridgeNode final
+/// \brief UdpSenderNode class which can send UDP datagrams
+class UdpSenderNode final
   : public lc::LifecycleNode
 {
 public:
   /// \brief Default constructor
   /// \param[in] options Options for the node
-  explicit SerialBridgeNode(const rclcpp::NodeOptions & options);
+  explicit UdpSenderNode(const rclcpp::NodeOptions & options);
 
   /// \brief Constructor which accepts IoContext
   /// \param[in] options Options for the node
   /// \param[in] ctx A shared IoContext
-  SerialBridgeNode(
+  UdpSenderNode(
     const rclcpp::NodeOptions & options,
-    const IoContext & ctx);
+    const drivers::common::IoContext & ctx);
 
   /// \brief Destructor - required to manage owned IoContext
-  ~SerialBridgeNode();
+  ~UdpSenderNode();
 
   /// \brief Callback from transition to "configuring" state.
   /// \param[in] state The current state that the node is in.
@@ -76,24 +74,20 @@ public:
   /// \param[in] state The current state that the node is in.
   LNI::CallbackReturn on_shutdown(const lc::State & state) override;
 
-  /// \brief Callback for sending a raw serial message
-  void subscriber_callback(const UInt8MultiArray::SharedPtr msg);
-
-  /// \breif Callback for when serial data are received
-  void receive_callback(const std::vector<uint8_t> & buffer, const size_t & bytes_transferred);
+  /// \brief Callback for sending a UDP datagram
+  void subscriber_callback(udp_msgs::msg::UdpPacket::SharedPtr msg);
 
 private:
   void get_params();
 
-  std::unique_ptr<IoContext> m_owned_ctx{};
-  std::string m_device_name{};
-  std::unique_ptr<SerialPortConfig> m_device_config;
-  std::unique_ptr<SerialDriver> m_serial_driver;
-  lc::LifecyclePublisher<UInt8MultiArray>::SharedPtr m_publisher;
-  rclcpp::Subscription<UInt8MultiArray>::SharedPtr m_subscriber;
-};  // class SerialBridgeNode
+  std::unique_ptr<drivers::common::IoContext> m_owned_ctx{};
+  std::string m_ip{};
+  uint16_t m_port{};
+  std::unique_ptr<UdpDriver> m_udp_driver;
+  rclcpp::Subscription<udp_msgs::msg::UdpPacket>::SharedPtr m_subscriber;
+};  // class UdpSenderNode
 
-}  // namespace serial_driver
+}  // namespace udp_driver
 }  // namespace drivers
 
-#endif  // SERIAL_DRIVER__SERIAL_BRIDGE_NODE_HPP_
+#endif  // UDP_DRIVER__UDP_SENDER_NODE_HPP_
