+++ udp_driver/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>udp_driver</name>
+  <name>boost_udp_driver</name>
   <version>1.2.0</version>
   <description>A library to write Synchronous and Asynchronous networking applications, ROS and ROS2 nodes</description>
   <maintainer email="whitleysoftwareservices@gmail.com">Joshua Whitley</maintainer>
@@ -13,7 +13,7 @@
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
 
   <depend>boost</depend>
-  <depend>io_context</depend>
+  <depend>boost_io_context</depend>
   <depend>lifecycle_msgs</depend>
   <depend>rclcpp</depend>
   <depend>rclcpp_components</depend>
