+++ udp_driver/test/test_udp_data.cpp
@@ -18,7 +18,7 @@
 #include <thread>
 #include <vector>
 
-#include "udp_driver/udp_socket.hpp"
+#include "boost_udp_driver/udp_socket.hpp"
 
 using drivers::common::IoContext;
 using drivers::udp_driver::UdpSocket;
