+++ udp_driver/test/test_udp_driver.cpp
@@ -22,8 +22,8 @@
 
 #include <rclcpp/rclcpp.hpp>
 
-#include "io_context/io_context.hpp"
-#include "udp_driver/udp_driver.hpp"
+#include "boost_io_context/io_context.hpp"
+#include "boost_udp_driver/udp_driver.hpp"
 
 using drivers::common::IoContext;
 using drivers::udp_driver::UdpDriver;
